Implemented Shell32TreeView and updated URIs in FileManager
In this commit, we have implemented a new `TreeView` control, `Shell32TreeView`, in `Shell32TreeView.xaml` and `Shell32TreeView.xaml.cs`. This control uses a `DataTemplate` to define the display of each item and a new class `Shell32TreeViewItem` to represent each item. The `RootShellItems` property was added to the `Shell32TreeView` class and initialized in the constructor.

We have also updated the URIs for `BitmapImage` in `DosShellItem.cs` and `DosShellItemHelpers.cs` to use `"ShellItem"` instead of `"Folder"`. The constants `shell32DefaultFolderIcon` and `shell32FolderContainingFileIcon` were updated accordingly, and the methods `GetIconPath` now return the new URIs.

In `FileManagerViewModel.cs`, the calls to `ShellTreeViewItems_GetItemsAsync` and `ShellGridViewItems_GetItemsAsync` were commented out.

In `FileManagerPage.xaml`, the `TreeView` in the `NavigationBarBorder` was replaced with a `Shell32TreeView` control and the `xmlns:vanara` namespace was added. The line setting the `ItemsSource` of `ShellTreeView` was removed in `FileManagerPage.xaml.cs`.