`TreeViewItem` element updated with new attributes
Updated the `TreeViewItem` element in several ways. The `Content` attribute, previously bound to `DisplayName`, has been removed, meaning the element no longer displays the `DisplayName` as its content. An `IsExpanded` attribute has been added with a default value of `True`, causing the element to always be expanded. Additionally, an `ItemsSource` attribute has been added and bound to `Children`, allowing the element to display child items if any are present.