Refactored Shell32GridView and Shell32TreeView classes
Refactored the `Shell32GridView` class in `Shell32GridView.xaml.cs` to initialize `GridShellItems` with a new `ObservableCollection<Shell32GridViewItem>` and `CurrentFolder` with a new `ShellFolder` instance pointing to the `c:\` directory by default. Simplified the anonymous method by removing the try-catch block and directly retrieving and adding `parentItem` to the `items` collection.

Modified the commented out code in `Shell32TreeView` class in `Shell32TreeView.xaml.cs` that adds new `Shell32TreeViewItem` instances to the `rootShellItem.Children` collection.

Refactored the `EnumerateChildren` method in `Shell32TreeViewItem` class to use pattern matching to cast `ShellItem` to `ShellFolder` and call `EnumerateChildren(filter)` on it. If the cast is not successful, an empty `Enumerable<ShellItem>` is returned.