Refactor EnumerateItems method and update Navigate method
Refactored the EnumerateItems method by updating the 'using' statement to the new syntax and adding a null check for 'newTarget'. The code that adds a Shell32GridViewItem for each child of 'newTarget' has been moved outside the 'if' statement. The 'return items;' statement has been moved to the end of the method. Also, the code calling EnumerateItems in the Navigate method has been commented out.