Refactor Shell32GridView and FileManagerViewModel classes
In Shell32GridView.xaml.cs, the CurrentFolder property has been refactored to use the Navigate function. The OnItemClickHandler method has been updated to return immediately if e.ClickedItem is not a Shell32GridViewItem item, and the nested condition has been moved out of the outer condition. In FileManagerViewModel.cs, the ShellTreeView and ShellGridView properties, along with the TreeView_OnItemInvoked method, have been commented out and moved to an obsolete code region.